+++ CHANGELOG.md
@@ -1,10 +1,22 @@
 # Changelog
 
-## v1.0.2
+## v1.1.0-0
+
+### 🚀 Enhancements
+
+- **types:** Added module types (10ec705)
+- Added support to include uikit default css with options (e9b916c)
+
+### 🩹 Fixes
+
+- Reset versioning due a typo in the version while releasing (d6079c1)
 
 ### 📖 Documentation
 
 - Updated README.md (2265556)
+- Added accordion test/example (1dbb616)
+- Add usage docs (84efbfd)
+- Usage instruction and examples (8df71b2)
 
 ### 🏡 Chore
 
@@ -12,6 +24,10 @@
 - Changed husky permissions to executable (c75c6b6)
 - **dev-deps:** Added uikit && types/uikit (cecc7e1)
 - First dev release (40c30ee)
+- **version-bump:** Version bump when released 1.0.2-dev (da6ba6a)
+- Changed uikit from devDep to peerDep (3d97e1c)
+- Added accordion component to test in Nuxt Playground (0df3432)
+- Updated package.json (1689fba)
 
 ### ❤️ Contributors